Moved SelectorUtils into internal.query.Helpers
