+++ cloudant-client/src/main/java/com/cloudant/client/internal/query/Builder.java
@@ -18,7 +18,6 @@ import com.cloudant.client.api.Database;
 import com.cloudant.client.api.query.Field;
 import com.cloudant.client.api.query.Selector;
 import com.cloudant.client.api.query.TextIndex;
-import com.cloudant.client.internal.util.SelectorUtils;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 
@@ -71,7 +70,7 @@ public abstract class Builder<I extends InternalIndex<D, F>, D extends Definitio
      * @return the builder for chaining
      */
     public B partialFilterSelector(Selector selector) {
-        instance.def.selector = SelectorUtils.getJsonObjectFromSelector(selector);
+        instance.def.selector = Helpers.getJsonObjectFromSelector(selector);
         return returnThis();
     }
